% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ajmrzik16_col}
\alias{ajmrzik16_col}
\title{Zika cases by municipality in the departament of Tolima, Colombia.}
\format{A sf object with 46 rows and 6 variables:
\describe{
  \item{NAME_0}{the country.}
  \item{NAME_1}{the name of departament.}
  \item{NAME_2}{the name of municipality.}
  \item{ZIKV_1516}{number of zika cases in 2015 & 2016 year.}
  \item{geometry}{simple feature.}
  ...
}}
\source{
\url{https://f1000research.com/articles/5-568/v1#DS0}
}
\usage{
ajmrzik16_col
}
\description{
A dataset containing the zika cases by AGEB in the years 2015 & 2016
Tolima departament.
}
\keyword{datasets}
