% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{den_loc_mex}
\alias{den_loc_mex}
\title{Dengue cases by AGEBs from nine locations in Mexico.}
\format{A sf object with 2174 rows and 12 variables:
\describe{
  \item{CVEGEO}{the ids }
  \item{loc}{the name of locality}
  \item{DENV_2008}{number of dengue cases in 2008.}
  \item{DENV_2009}{number of dengue cases in 2009.}
  \item{DENV_2010}{number of dengue cases in 2010.}
  \item{DENV_2011}{number of dengue cases in 2011.}
  \item{DENV_2012}{number of dengue cases in 2012.}
  \item{DENV_2013}{number of dengue cases in 2013.}
  \item{DENV_2014}{number of dengue cases in 2014.}
  \item{DENV_2015}{number of dengue cases in 2015.}
  \item{DENV_2016}{number of dengue cases in 2016.}
  \item{geometry}{simple feature.}
  ...
}}
\source{
\url{https://osf.io/8bjy5/}
}
\usage{
den_loc_mex
}
\description{
A dataset containing the clinical dengue cases by AGEB and year (2008-2016)
from nine location of Mexico.
}
\keyword{datasets}
