% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{den_loc_mex}
\alias{den_loc_mex}
\title{Dengue cases by AGEBs from nine locations in Mexico.}
\format{
A sf object with 2174 rows and 12 variables:
\describe{
  \item{loc}{the name of locality}
  \item{CVEGEO}{the ids }
  \item{DENV2008}{number of dengue cases in 2008.}
  \item{DENV2009}{number of dengue cases in 2009.}
  \item{DENV2010}{number of dengue cases in 2010.}
  \item{DENV2011}{number of dengue cases in 2011.}
  \item{DENV2012}{number of dengue cases in 2012.}
  \item{DENV2013}{number of dengue cases in 2013.}
  \item{DENV2014}{number of dengue cases in 2014.}
  \item{DENV2015}{number of dengue cases in 2015.}
  \item{DENV2016}{number of dengue cases in 2016.}
  \item{CHIKV2015}{number of chikungunya cases in 2015.}
  \item{CHIKV2016}{number of chikungunya cases in 2016.}
  \item{ZIKV2016}{number of zika cases in 2016.}
  \item{geometry}{simple feature.}
  ...
}
}
\source{
\url{https://data.mendeley.com/datasets/m6nvnzgzdt/1}
}
\usage{
den_loc_mex
}
\description{
A dataset containing the clinical dengue cases by AGEB and year (2008-2016)
from nine location of Mexico.
}
\keyword{datasets}
